+++ src/api/deleteCartItem.ts
@@ -1,0 +1,8 @@
+import { baseAPI } from './baseAPI';
+
+export async function deleteCartItem(cartId: string) {
+  await baseAPI({
+    method: 'DELETE',
+    path: `/cart-items/${cartId}`,
+  });
+}
