+++ src/components/features/product/product-list/ProductList.tsx
@@ -19,6 +19,7 @@ function ProductList({
           <ProductCard
             key={id}
             id={id}
+            data-testid={id}
             cartId={matchingCart?.id ?? null}
             name={name}
             price={price}
