+++ src/shop/__mocks__/productListMockData.json
@@ -19,40 +19,5 @@
     "price": 2100,
     "imageUrl": "https://sitem.ssgcdn.com/88/19/87/item/0000006871988_i1_750.jpg",
     "category": "식료품"
-  },
-  {
-    "id": 28,
-    "name": "아샷추",
-    "price": 3800,
-    "imageUrl": "https://d2afncas1tel3t.cloudfront.net/wp-content/uploads/2023/12/%EC%95%84%EC%83%B7%EC%B6%94%EC%95%84%EC%9D%B4%EC%8A%A4%ED%8B%B0%EC%83%B7%EC%B6%94%EC%B9%B4%EB%94%94%EC%B9%B4%ED%8E%98%EC%9D%B8_1.png",
-    "category": "식료품"
-  },
-  {
-    "id": 27,
-    "name": "아바라",
-    "price": 4800,
-    "imageUrl": "https://image.ohousecdn.com/i/bucketplace-v2-development/uploads/cards/snapshots/171653801239329270.jpeg?w=256&h=366&c=c",
-    "category": "식료품"
-  },
-  {
-    "id": 25,
-    "name": "얌샘김밥",
-    "price": 5000,
-    "imageUrl": "https://search.pstatic.net/common/?src=https%3A%2F%2Fldb-phinf.pstatic.net%2F20171018_6%2F1508253136417Dlrjh_PNG%2FCdq22zpVpr92_XHROlHbxjJ0.png&type=sc960_832",
-    "category": "식료품"
-  },
-  {
-    "id": 6,
-    "name": "플라망고",
-    "price": 8130,
-    "imageUrl": "https://velog.velcdn.com/images/minsungje/post/c27c57cb-fcbb-4641-b72d-0e2030739ae7/image.jpg",
-    "category": "식료품"
-  },
-  {
-    "id": 5,
-    "name": "동물 양말",
-    "price": 20000,
-    "imageUrl": "https://m.cocosocks.com/web/product/medium/202503/940897aced51144109baa4d145def01f.jpg",
-    "category": "패션잡화"
   }
 ]
